Handle start and end dates smartly
Adds two new options `start` and `end` to the `read` command. The
default to `start` is the current date, and the default to `end` is the
last day of the current year. The last day of the current semester would
be better, but this seems functionally equivalent since no semester
calendar stretches into the following year.